blog/requirements.txt
aiohttp==3.5.4
amqp==2.4.1
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bleach==3.1.0
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.5
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.4
django-haystack==2.8.1
django-redis==4.10.0
django-redis-sessions==0.6.1
django-tinymce==2.8.0
dnspython==1.16.0
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.24.1
fake-useragent==0.1.11
Flask==1.0.2
greenlet==0.4.15
html-testRunner==1.1.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
jieba==0.39
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.3.0
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
monotonic==1.5
multidict==4.5.2
mysql-connector==2.1.6
mysql-connector-python==8.0.13
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
openpyxl==2.5.14
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
pyee==5.0.0
pygame==1.9.4
Pygments==2.3.1
PyHamcrest==1.9.0
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.3.0
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
pytz==2018.7
pywin32==224
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
redis==3.0.1
requests==2.21.0
retrying==1.3.3
scipy==1.2.0
Scrapy==1.5.2
seaborn==0.9.0
selenium==2.48.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.6.2
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.24.1
views==0.3
vine==1.2.0
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
Werkzeug==0.14.1
Whoosh==2.7.4
widgetsnbextension==3.4.2
xlrd==1.2.0
yarl==1.3.0
zope.interface==4.6.0
